requirements.txt
APScheduler==3.11.0
Flask==3.1.1
Flask-Login==0.6.3
mysql-connector-python==8.3.0
openai==1.95.1
pandas==2.3.1
python-dotenv==1.1.1
Requests==2.32.4
Werkzeug==3.1.3
gunicorn
# xhtml2pdf==0.2.17

